Code updates:
Added new code to Pipe1BrokenScript that corresponds to launching the broken pipe child according to the collision.
Created an OnCollisionEnter function that applies a AddForceAtPosition where the Bord hit the broken pipe.
Added flags to make sure force is only applied one time to avoid funny behaviour.

Also made the Bord have a trigger collider and a physics collider. Both are child of the GameObject.

Next time, need to add a broken pipe base script to continue the movement of the pipe base.
Also need a despawning system for the broken pipes.